// frontend/src/app/models/mojDogadjaj.ts
import { Osoba } from "./osoba"

export class MojDogadjaj{
    naziv: string;
    mesto: string;
    vreme: Date;
    limit: Date;
    opis: string;
    osobe: Osoba[];
    hrana: string[];
    pice:  string[];
    ostalo: string[];
    zelje:  string[];
    ulica: string;
    korisnickoIme: string;
    id: number;
    idMesta: number;

}